data_shift_service/requirements.txt
# Core dependencies
google-cloud-logging==3.8.0
google-cloud-bigquery==3.11.4
requests==2.31.0
APScheduler==3.10.4
pytz==2023.3

# Data processing
pandas==2.0.3
numpy==1.24.3

# Async support
aiohttp==3.8.5
asyncio-mqtt==0.13.0

# Email support
secure-smtplib==0.1.1

# Utilities
python-dateutil==2.8.2
pydantic==2.4.2
